+++ src/components/layout/header/About.tsx
@@ -1,3 +1,5 @@
+import { useEffect, useState } from 'react';
+
 import {
     Button,
     HStack,
@@ -11,10 +13,20 @@ import {
     ModalOverlay,
     useDisclosure,
     VStack,
+    Text,
+    Spinner,
+    Link,
+    Code,
+    Box,
+    Alert,
+    AlertIcon,
+    AlertDescription,
 } from '@chakra-ui/react';
 
 import useAuthStore from '@/features/auth/useAuthStore';
 import CopyButton from '@/features/copy/CopyButton';
+import useReleasesInfoStore from '@/features/docs/useReleasesInfoStore';
+import { ReleaseInfoT, ReleasesInfoT } from '@/types/DocsTypes';
 
 import { DrawerSection, DrawerSectionEntry } from '@/components/DrawerSection';
 
@@ -28,6 +40,56 @@ const About = () => {
             chaincode_version: chaincodeVersion,
         },
     } = useAuthStore();
+
+    const {
+        info: releasesInfo,
+        fetchingInfo: fetchingReleasesInfo,
+        fetchInfo: fetchReleasesInfo,
+    } = useReleasesInfoStore();
+    const [matchingRelease, setMatchingRelease] =
+        useState<ReleaseInfoT | null>();
+
+    useEffect(() => {
+        if (isOpen && releasesInfo === null) fetchReleasesInfo();
+    }, [isOpen, fetchReleasesInfo, releasesInfo]);
+
+    useEffect(() => {
+        if (releasesInfo && backendVersion && orchestratorVersion) {
+            setMatchingRelease(
+                getMatchingSubstraRelease(
+                    __APP_VERSION__,
+                    backendVersion,
+                    orchestratorVersion,
+                    releasesInfo
+                )
+            );
+        } else {
+            setMatchingRelease(null);
+        }
+    }, [releasesInfo, backendVersion, orchestratorVersion]);
+
+    const getMatchingSubstraRelease = (
+        frontendVersion: string,
+        backendVersion: string,
+        orchestratorVersion: string,
+        releases: ReleasesInfoT
+    ) => {
+        const candidateReleases = releases.releases.filter((it) => {
+            return (
+                it.components['substra-frontend'].version ===
+                    frontendVersion.split('+')[0] &&
+                it.components['substra-backend'].version ===
+                    backendVersion.split('+')[0] &&
+                it.components.orchestrator.version ===
+                    orchestratorVersion.split('+')[0]
+            );
+        });
+        if (candidateReleases.length === 0) {
+            return null;
+        } else {
+            return candidateReleases[0];
+        }
+    };
 
     return (
         <>
@@ -58,6 +120,83 @@ const About = () => {
                                     </DrawerSectionEntry>
                                 )}
                             </DrawerSection>
+                            <Box fontSize="xs">
+                                {fetchingReleasesInfo ? (
+                                    <HStack>
+                                        <Spinner
+                                            color="gray.400"
+                                            size="xs"
+                                            label="Fetching bundle info"
+                                        />
+                                        <Text>{'Fetching bundle info'}</Text>
+                                    </HStack>
+                                ) : (
+                                    <>
+                                        {matchingRelease ? (
+                                            <Alert
+                                                status="info"
+                                                alignItems={'center'}
+                                            >
+                                                <AlertIcon />
+                                                <AlertDescription>
+                                                    <Text>
+                                                        {
+                                                            'This server is running Substra '
+                                                        }
+                                                        <Text as="b">
+                                                            {`${matchingRelease.version}`}
+                                                        </Text>
+                                                        {'.'}
+                                                    </Text>
+                                                    <Text>
+                                                        {
+                                                            'In Python, you should use '
+                                                        }
+                                                        <Code fontSize="xs">
+                                                            {`substra==${matchingRelease.components.substra.version}`}
+                                                        </Code>
+                                                        {' and/or '}
+                                                        <Code fontSize="xs">
+                                                            {`substrafl==${matchingRelease.components.substrafl.version}`}
+                                                        </Code>
+                                                        {'.'}
+                                                    </Text>
+                                                </AlertDescription>
+                                            </Alert>
+                                        ) : (
+                                            <Alert
+                                                status="warning"
+                                                alignItems={'center'}
+                                            >
+                                                <AlertIcon />
+                                                <AlertDescription>
+                                                    <Text>
+                                                        {
+                                                            'This server is not running a '
+                                                        }
+
+                                                        <Link
+                                                            color="black"
+                                                            href="https://docs.substra.org/en/latest/additional/release.html"
+                                                            isExternal
+                                                            textDecoration={
+                                                                'underline'
+                                                            }
+                                                            textUnderlineOffset={
+                                                                3
+                                                            }
+                                                        >
+                                                            known Substra
+                                                            release
+                                                        </Link>
+                                                        {'.'}
+                                                    </Text>
+                                                </AlertDescription>
+                                            </Alert>
+                                        )}
+                                    </>
+                                )}
+                            </Box>
                             <DrawerSection title="Backend">
                                 <DrawerSectionEntry title="URL">
                                     <HStack spacing={1.5}>