+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Substra bundle version info in the "About" modal (#256)
+
 ## [0.45.1] - 2023-10-06
 
 ### Changed